Move zig installation to its own module.
